+++ riscv_isac/__init__.py
@@ -3,5 +3,5 @@
 """Top-level package for RISC-V ISA Coverage."""
 
 __author__ = """InCore Semiconductors Pvt Ltd"""
-__email__ = 'incorebot@gmail.com'
-__version__ = '0.1.0'
+__email__ = 'info@incoresemi.com'
+__version__ = '0.2.0'