+++ riscv_isac/parsers.py
@@ -127,7 +127,7 @@ def parseInstruction(input_line, mode, arch='rv32'):
     first_two_bits = FIRST2_MASK & instr
 
     if first_two_bits == 0b11:
-        return parseStandardInstruction(input_line, mode)
+        return parseStandardInstruction(input_line, mode, arch)
     else:
         return parseCompressedInstruction(input_line, mode, arch)
 
@@ -152,7 +152,7 @@ def parseCompressedInstruction(input_line, mode, arch):
 
     return instrObj
 
-def parseStandardInstruction(input_line, mode):
+def parseStandardInstruction(input_line, mode, arch):
     ''' Parse an input line and decode the instruction
         Args: input_line - Line from the log file
         Returns: (instr_name, rd, rs1, rs2, imm)
@@ -165,7 +165,7 @@ def parseStandardInstruction(input_line, mode):
     addr = extractAddress(input_line, mode)
     opcode = extractOpcode(instr)
     try:
-        instrObj = opcodes[opcode](instr, addr)
+        instrObj = opcodes[opcode](instr, addr, arch)
     except KeyError as e:
         print("Instruction not found", hex(instr))
         return None
@@ -178,17 +178,17 @@ def twos_comp(val, bits):
         val = val - (1 << bits)
     return val
 
-def lui(instr, addr):
+def lui(instr, addr, arch):
     imm = instr >> 12
     rd = ((instr & RD_MASK) >> 7, 'x')
     return instructionObject('lui', addr, rd = rd, imm = imm)
 
-def auipc(instr, addr):
+def auipc(instr, addr, arch):
     imm = instr >> 12
     rd = ((instr & RD_MASK) >> 7, 'x')
     return instructionObject('auipc', addr, rd = rd, imm = imm)
 
-def jal(instr, addr):
+def jal(instr, addr, arch):
     imm_10_1 = (instr >> 21) & 0x000003ff
     imm_11 = (instr >> 10) & 0x00000400
     imm_19_12 = (instr & 0x000ff000) >> 1
@@ -198,14 +198,14 @@ def jal(instr, addr):
     rd = ((instr & RD_MASK) >> 7, 'x')
     return instructionObject('jal', addr, rd = rd, imm = imm)
 
-def jalr(instr, addr):
+def jalr(instr, addr, arch):
     rd = ((instr & RD_MASK) >> 7, 'x')
     rs1 = ((instr & RS1_MASK) >> 15, 'x')
     imm_11_0 = instr >> 20
     imm = twos_comp(imm_11_0, 12)
     return instructionObject('jalr', addr, rd = rd, rs1 = rs1, imm = imm)
 
-def branch_ops(instr, addr):
+def branch_ops(instr, addr, arch):
     funct3 = (instr & FUNCT3_MASK) >> 12
     rs1 = ((instr & RS1_MASK) >> 15, 'x')
     rs2 = ((instr & RS2_MASK) >> 20, 'x')
@@ -233,7 +233,7 @@ def branch_ops(instr, addr):
 
     return instrObj
 
-def load_ops(instr, addr):
+def load_ops(instr, addr, arch):
     funct3 = (instr & FUNCT3_MASK) >> 12
     rd = ((instr & RD_MASK) >> 7, 'x')
     rs1 = ((instr & RS1_MASK) >> 15, 'x')
@@ -258,7 +258,7 @@ def load_ops(instr, addr):
 
     return instrObj
 
-def store_ops(instr, addr):
+def store_ops(instr, addr, arch):
     funct3 = (instr & FUNCT3_MASK) >> 12
     rs1 = ((instr & RS1_MASK) >> 15, 'x')
     rs2 = ((instr & RS2_MASK) >> 20, 'x')
@@ -279,7 +279,7 @@ def store_ops(instr, addr):
 
     return instrObj
 
-def arithi_ops(instr, addr):
+def arithi_ops(instr, addr, arch):
     funct3 = (instr & FUNCT3_MASK) >> 12
     rd = ((instr & RD_MASK) >> 7, 'x')
     rs1 = ((instr & RS1_MASK) >> 15, 'x')
@@ -304,11 +304,17 @@ def arithi_ops(instr, addr):
     if funct3 == 0b001:
         instrObj.instr_name = 'slli'
         instrObj.imm = None
-        shamt = imm & 0x01f
+        if arch == 'rv32':
+            shamt = imm & 0x01f
+        elif arch == 'rv64':
+            shamt = imm & 0x03f
         instrObj.shamt = shamt
     if funct3 == 0b101:
         instrObj.imm = None
-        shamt = imm & 0x01f
+        if arch == 'rv32':
+            shamt = imm & 0x01f
+        elif arch == 'rv64':
+            shamt = imm & 0x03f
         instrObj.shamt = shamt
         rtype_bit = (imm >> 10) & 0x1
         if rtype_bit == 1:
@@ -345,7 +351,7 @@ def arithm_ops(instr, addr):
 
     return instrObj
 
-def arith_ops(instr, addr):
+def arith_ops(instr, addr, arch):
 
     # Test for RV32M ops
     funct7 = (instr >> 25)
@@ -388,7 +394,7 @@ def arith_ops(instr, addr):
 
     return instrObj
 
-def fence_ops(instr, addr):
+def fence_ops(instr, addr, arch):
     funct3 = (instr & FUNCT3_MASK) >> 12
 
     pred = (instr >> 20) & 0x0000000f
@@ -405,7 +411,7 @@ def fence_ops(instr, addr):
 
     return instrObj
 
-def control_ops(instr, addr):
+def control_ops(instr, addr, arch):
     funct3 = (instr & FUNCT3_MASK) >> 12
 
     # Test for ecall and ebreak ops
@@ -444,7 +450,7 @@ def control_ops(instr, addr):
 
     return instrObj
 
-def rv64i_arithi_ops(instr, addr):
+def rv64i_arithi_ops(instr, addr, arch):
     funct3 = (instr & FUNCT3_MASK) >> 12
     rd = ((instr & RD_MASK) >> 7, 'x')
     rs1 = ((instr & RS1_MASK) >> 15, 'x')
@@ -477,20 +483,20 @@ def rv64m_arithm_ops(instr, addr):
     instrObj = instructionObject('None', addr, rd = rd, rs1 = rs1, rs2 = rs2)
 
     if funct3 == 0b000:
-        instrObj.instr_name == 'mulw'
+        instrObj.instr_name = 'mulw'
     if funct3 == 0b100:
-        instrObj.instr_name == 'divw'
+        instrObj.instr_name = 'divw'
     if funct3 == 0b101:
-        instrObj.instr_name == 'divuw'
+        instrObj.instr_name = 'divuw'
     if funct3 == 0b110:
-        instrObj.instr_name == 'remw'
+        instrObj.instr_name = 'remw'
     if funct3 == 0b111:
-        instrObj.instr_name == 'remuw'
+        instrObj.instr_name = 'remuw'
 
     return instrObj
 
 
-def rv64i_arith_ops(instr, addr):
+def rv64i_arith_ops(instr, addr, arch):
 
     # Test for rv64m ops
     funct7 = (instr >> 25)
@@ -550,7 +556,7 @@ rv64a_instr_names = {
         0b11100: 'amomaxu.d'
 }
 
-def rv64_rv32_atomic_ops(instr, addr):
+def rv64_rv32_atomic_ops(instr, addr, arch):
     funct5 = (instr >> 27) & 0x0000001f
     funct3 = (instr & FUNCT3_MASK) >> 12
     rd = ((instr & RD_MASK) >> 7, 'x')
@@ -583,7 +589,7 @@ def rv64_rv32_atomic_ops(instr, addr):
 
         return instrObj
 
-def flw_fld(instr, addr):
+def flw_fld(instr, addr, arch):
     rd = ((instr & RD_MASK) >> 7, 'f')
     rs1 = ((instr & RS1_MASK) >> 15, 'x')
     funct3 = (instr & FUNCT3_MASK) >> 12
@@ -598,7 +604,7 @@ def flw_fld(instr, addr):
 
     return instrObj
 
-def fsw_fsd(instr, addr):
+def fsw_fsd(instr, addr, arch):
     imm_4_0 = (instr & RD_MASK) >> 7
     imm_11_5 = (instr >> 25) << 5
     imm = twos_comp(imm_4_0 + imm_11_5, 12)
@@ -616,7 +622,7 @@ def fsw_fsd(instr, addr):
 
     return instrObj
 
-def fmadd(instr, addr):
+def fmadd(instr, addr, arch):
     rd = ((instr & RD_MASK) >> 7, 'f')
     rm = (instr >> 12) & 0x00000007
     rs1 = ((instr & RS1_MASK) >> 15, 'f')
@@ -635,7 +641,7 @@ def fmadd(instr, addr):
 
     return instrObj
 
-def fmsub(instr, addr):
+def fmsub(instr, addr, arch):
     rd = ((instr & RD_MASK) >> 7, 'f')
     rm = (instr >> 12) & 0x00000007
     rs1 = ((instr & RS1_MASK) >> 15, 'f')
@@ -654,7 +660,7 @@ def fmsub(instr, addr):
 
     return instrObj
 
-def fnmsub(instr, addr):
+def fnmsub(instr, addr, arch):
     rd = ((instr & RD_MASK) >> 7, 'f')
     rm = (instr >> 12) & 0x00000007
     rs1 = ((instr & RS1_MASK) >> 15, 'f')
@@ -673,7 +679,7 @@ def fnmsub(instr, addr):
 
     return instrObj
 
-def fnmadd(instr, addr):
+def fnmadd(instr, addr, arch):
     rd = ((instr & RD_MASK) >> 7, 'f')
     rm = (instr >> 12) & 0x00000007
     rs1 = ((instr & RS1_MASK) >> 15, 'f')
@@ -692,7 +698,7 @@ def fnmadd(instr, addr):
 
     return instrObj
 
-def rv32_rv64_float_ops(instr, addr):
+def rv32_rv64_float_ops(instr, addr, arch):
     rd = ((instr & RD_MASK) >> 7, 'f')
     rm = (instr >> 12) & 0x00000007
     rs1 = ((instr & RS1_MASK) >> 15, 'f')
@@ -1055,7 +1061,7 @@ def quad1(instr, addr, arch):
     imm_4_0 = (C1_IMM_4_0_MASK & instr) >> 2
     imm = imm_5 + imm_4_0
 
-    imm_lui = ((C1_IMM_17_MASK & instr) << 5) + ((C1_IMM_16_12_MASK & instr) << 10)
+    imm_lui = ((C1_IMM_17_MASK & instr) >> 12) + ((C1_IMM_16_12_MASK & instr)>>2)
     imm_j_5 = get_bit(instr, 2) << 5
     imm_j_3_1 = get_bit(instr,3) + (get_bit(instr, 4) << 1) + (get_bit(instr,5) << 2)
     imm_j_7 = get_bit(instr,6) << 7
@@ -1076,7 +1082,7 @@ def quad1(instr, addr, arch):
     imm_addi_5 = get_bit(instr, 2) << 5
     imm_addi_8_7 = (get_bit(instr, 3) + (get_bit(instr, 4)<< 1) ) << 7
     imm_addi_6 = get_bit(instr, 5) << 6
-    imm_addi_4 = get_bit(instr, 8) << 4
+    imm_addi_4 = get_bit(instr, 6) << 4
     imm_addi_9 = get_bit(instr, 12) << 9
     imm_addi = imm_addi_5 + imm_addi_8_7 + imm_addi_6 + imm_addi_4 + imm_addi_9
 
@@ -1087,7 +1093,7 @@ def quad1(instr, addr, arch):
         instrObj.rs1 = (rs1, 'x')
         instrObj.rd = (rd, 'x')
         instrObj.imm = twos_comp(imm, 6)
-        if imm == 0 and rd == 0:
+        if rd == 0:
             instrObj.instr_name = 'c.nop'
         else:
             instrObj.instr_name = 'c.addi'
@@ -1101,7 +1107,7 @@ def quad1(instr, addr, arch):
             instrObj.imm = twos_comp(imm, 6)
             instrObj.rs1 = (rs1, 'x')
             instrObj.rd = (rd, 'x')
-    elif funct3 == 2 and rd != 0:
+    elif funct3 == 2:
         instrObj.instr_name = 'c.li'
         instrObj.imm = twos_comp(imm, 6)
         instrObj.rd = (rd, 'x')
@@ -1110,10 +1116,11 @@ def quad1(instr, addr, arch):
             instrObj.instr_name = 'c.addi16sp'
             instrObj.rs1 = (2, 'x')
             instrObj.rd = (2, 'x')
-            instrObj.imm = twos_comp(imm_addi, 2)
-        elif rd !=0 and rd != 2 and imm_lui !=0:
+            instrObj.imm = twos_comp(imm_addi, 10)
+        elif rd != 2 and imm_lui !=0:
             instrObj.instr_name = 'c.lui'
-            instrObj.imm = twos_comp(imm_lui, 18)
+            instrObj.imm = imm
+            instrObj.rs1 = (rd, 'x')
             instrObj.rd = (rd, 'x')
     elif funct3 == 4:
         if op == 0 and imm != 0:
@@ -1168,12 +1175,12 @@ def quad1(instr, addr, arch):
     elif funct3 == 6:
         instrObj.instr_name = 'c.beqz'
         instrObj.rs1 = (8 + rs1prime, 'x')
-        instrObj.rs2 = 0
+        instrObj.rs2 = (0 , 'x')
         instrObj.imm = twos_comp(imm_b, 9)
     elif funct3 == 7:
         instrObj.instr_name = 'c.bnez'
         instrObj.rs1 = (8 + rs1prime, 'x')
-        instrObj.rs2 = 0
+        instrObj.rs2 = (0, 'x')
         instrObj.imm = twos_comp(imm_b, 9)
 
 
@@ -1215,7 +1222,6 @@ def quad2(instr, addr, arch):
     imm_fsdsp = imm_5_3 + imm_s_8_6
     imm_swsp = imm_5_2 + imm_s_7_6
 
-
     if funct3 == 0 and imm_slli !=0 and rd !=0:
         instrObj.instr_name = 'c.slli'
         instrObj.rd = (rd, 'x')
@@ -1237,7 +1243,7 @@ def quad2(instr, addr, arch):
         instrObj.rs1 = (2, 'x')
         instrObj.imm = imm_lwsp
     elif funct3 == 3 and arch == 'rv64':
-        instrObj.instr_name = 'c.fldsp'
+        instrObj.instr_name = 'c.ldsp'
         instrObj.rd = (rd, 'f')
         instrObj.rs1 = (2, 'x')
         instrObj.imm = imm_ldsp
@@ -1245,18 +1251,17 @@ def quad2(instr, addr, arch):
         instrObj.instr_name = 'c.jr'
         instrObj.rs1 = (rs1, 'x')
         instrObj.imm = 0
-    elif funct3 == 4 and rd !=0 and rs2!=0 and imm_5 == 0:
+    elif funct3 == 4 and rs2!=0 and imm_5 == 0:
         instrObj.instr_name = 'c.mv'
-        instrObj.rs1 = (0, 'x')
         instrObj.rs2 = (rs2, 'x')
         instrObj.rd = (rd, 'x')
-    elif funct3 == 4 and rd ==0 and rs2 == 0 and imm_5 == 1:
+    elif funct3 == 4 and rd ==0 and rs2 == 0 and imm_5 == 32:
         instrObj.instr_name = 'c.ebreak'
-    elif funct3 == 4 and imm_5 == 1 and rs1 !=0 and rs2 == 0:
+    elif funct3 == 4 and imm_5 == 32 and rs1 !=0 and rs2 == 0:
         instrObj.instr_name = 'c.jalr'
         instrObj.rs1 = (rs1, 'x')
         instrObj.rd = (1, 'x')
-    elif funct3 == 4 and imm_5 == 1 and rs1 != 0 and rs2 !=0 :
+    elif funct3 == 4 and imm_5 == 32 and rs2 !=0 :
         instrObj.instr_name = 'c.add'
         instrObj.rs1 = (rs1, 'x')
         instrObj.rs2 = (rs2, 'x')
@@ -1281,7 +1286,6 @@ def quad2(instr, addr, arch):
         instrObj.rs2 = (rs2, 'x')
         instrObj.imm = imm_fsdsp
         instrObj.rs1 = (2 , 'x')
-
     return instrObj
 
 ''' Instruction Op-codes Dict '''