+++ CHANGELOG.rst
@@ -3,5 +3,15 @@ CHANGELOG
 
 This project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-[0.1.0] - 2020-06-25
+[0.2.0] - 2020-10-23
+  - Added documentation for CGF and usage
+  - Added normalization routine as cli
+  - Added abstract functions
+  - using click for cli
+  - adding parsers for sail and spike
+  - added support for filtering based on labels
+  - added merge-reports cli command
 
+
+[0.1.0] - 2020-06-25
+  - initial draft